restoring dt client + models behavior
