+++ sdk/translation/azure-ai-translation-document/azure/ai/translation/document/_patch.py
@@ -6,25 +6,257 @@
 
 Follow our quickstart for examples: https://aka.ms/azsdk/python/dpcodegen/python/customize
 """
-from typing import List
 
-__all__: List[str] = []  # Add all objects you want publicly available to users at this package level
-
-
-def patch_sdk():
-    """Do not remove from this file.
-
-    `patch_sdk` is a last resort escape hatch that allows you to do customizations
-    you can't accomplish using the techniques described in
-    https://aka.ms/azsdk/python/dpcodegen/python/customize
-    """
-# ------------------------------------
-# Copyright (c) Microsoft Corporation.
-# Licensed under the MIT License.
-# ------------------------------------
-
+from typing import Any, List, Union, overload, Optional, cast, Tuple, TypeVar, Dict
 from enum import Enum
+import json
+import datetime
+
 from azure.core import CaseInsensitiveEnumMeta
+from azure.core.tracing.decorator import distributed_trace
+from azure.core.paging import ItemPaged
+from azure.core.credentials import AzureKeyCredential, TokenCredential
+from azure.core.pipeline.policies import HttpLoggingPolicy
+from azure.core.exceptions import HttpResponseError, ODataV4Format
+from azure.core.pipeline import PipelineResponse
+from azure.core.rest import (
+    HttpResponse,
+    AsyncHttpResponse,
+    HttpRequest,
+)
+from azure.core.polling import LROPoller
+from azure.core.polling.base_polling import (
+    LROBasePolling,
+    OperationResourcePolling,
+    _is_empty,
+    _as_json,
+    BadResponse,
+    OperationFailed,
+    _raise_if_bad_http_status_and_method,
+)
+
+from .models import (
+    TranslationStatus,
+    DocumentStatus,
+    DocumentTranslationInput,
+    DocumentTranslationFileFormat,
+    StorageInputType,
+    TranslationGlossary,
+    TranslationTarget,
+    DocumentTranslationError,
+    BatchRequest as _BatchRequest,
+    SourceInput as _SourceInput,
+    TargetInput as _TargetInput,
+    DocumentFilter as _DocumentFilter,
+    StartTranslationDetails as _StartTranslationDetails,
+)
+from .models._models import (
+    TranslationStatus as _TranslationStatus,
+    DocumentStatus as _DocumentStatus,
+)
+from .models._patch import convert_status
+
+COGNITIVE_KEY_HEADER = "Ocp-Apim-Subscription-Key"
+POLLING_INTERVAL = 1
+
+ResponseType = Union[HttpResponse, AsyncHttpResponse]
+PipelineResponseType = PipelineResponse[HttpRequest, ResponseType]
+PollingReturnType_co = TypeVar("PollingReturnType_co", covariant=True)
+
+_FINISHED = frozenset(["succeeded", "cancelled", "cancelling", "failed"])
+_FAILED = frozenset(["validationfailed"])
+
+
+class DocumentTranslationLROPoller(LROPoller[PollingReturnType_co]):
+    """A custom poller implementation for Document Translation. Call `result()` on the poller to return
+    a pageable of :class:`~azure.ai.translation.document.DocumentStatus`."""
+
+    @property
+    def id(self) -> str:
+        """The ID for the translation operation
+
+        :return: The str ID for the translation operation.
+        :rtype: str
+        """
+        if self._polling_method._current_body:   # pylint: disable=protected-access
+            return self._polling_method._current_body.id   # pylint: disable=protected-access
+        return self._polling_method._get_id_from_headers()   # pylint: disable=protected-access
+
+    @property
+    def details(self) -> TranslationStatus:
+        """The details for the translation operation
+
+        :return: The details for the translation operation.
+        :rtype: ~azure.ai.translation.document.TranslationStatus
+        """
+        if self._polling_method._current_body:   # pylint: disable=protected-access
+            return TranslationStatus._from_generated(  # pylint: disable=protected-access
+                self._polling_method._current_body   # pylint: disable=protected-access
+            )
+        return TranslationStatus(id=self._polling_method._get_id_from_headers())   # pylint: disable=protected-access
+
+    @classmethod
+    def from_continuation_token(  # pylint: disable=docstring-missing-return,docstring-missing-param,docstring-missing-rtype
+        cls, polling_method, continuation_token, **kwargs: Any
+    ):
+        """
+        :meta private:
+        """
+        (
+            client,
+            initial_response,
+            deserialization_callback,
+        ) = polling_method.from_continuation_token(continuation_token, **kwargs)
+
+        return cls(client, initial_response, deserialization_callback, polling_method)
+
+
+class DocumentTranslationLROPollingMethod(LROBasePolling):
+    """A custom polling method implementation for Document Translation."""
+
+    def __init__(self, *args, **kwargs):
+        self._cont_token_response = kwargs.pop("cont_token_response")
+        super().__init__(*args, **kwargs)
+
+    @property
+    def _current_body(self) -> _TranslationStatus:
+        try:
+            return _TranslationStatus(self._pipeline_response.http_response.json())
+        except json.decoder.JSONDecodeError:
+            return _TranslationStatus()
+
+    def _get_id_from_headers(self) -> str:
+        return self._initial_response.http_response.headers["Operation-Location"].split("/batches/")[1].split("?api-version")[0]
+
+    def finished(self) -> bool:
+        """Is this polling finished?
+
+        :return: True/False for whether polling is complete.
+        :rtype: bool
+        """
+        return self._finished(self.status())
+
+    @staticmethod
+    def _finished(status) -> bool:
+        if hasattr(status, "value"):
+            status = status.value
+        return str(status).lower() in _FINISHED
+
+    @staticmethod
+    def _failed(status) -> bool:
+        if hasattr(status, "value"):
+            status = status.value
+        return str(status).lower() in _FAILED
+
+    def get_continuation_token(self) -> str:
+        if self._current_body:
+            return self._current_body.id
+        return self._get_id_from_headers()
+
+    # pylint: disable=arguments-differ
+    def from_continuation_token(self, continuation_token: str, **kwargs: Any) -> Tuple:  
+        try:
+            client = kwargs["client"]
+        except KeyError as exc:
+            raise ValueError("Need kwarg 'client' to be recreated from continuation_token") from exc
+
+        try:
+            deserialization_callback = kwargs["deserialization_callback"]
+        except KeyError as exc:
+            raise ValueError("Need kwarg 'deserialization_callback' to be recreated from continuation_token") from exc
+
+        return client, self._cont_token_response, deserialization_callback
+
+    def _poll(self) -> None:
+        """Poll status of operation so long as operation is incomplete and
+        we have an endpoint to query.
+
+        :raises: OperationFailed if operation status 'Failed' or 'Canceled'.
+        :raises: BadStatus if response status invalid.
+        :raises: BadResponse if response invalid.
+        """
+
+        while not self.finished():
+            self.update_status()
+        while not self.finished():
+            self._delay()
+            self.update_status()
+
+        if self._failed(self.status()):
+            raise OperationFailed("Operation failed or canceled")
+
+        final_get_url = self._operation.get_final_get_url(self._pipeline_response)  
+        if final_get_url:
+            self._pipeline_response = self.request_status(final_get_url)
+            _raise_if_bad_http_status_and_method(self._pipeline_response.http_response)
+
+
+class TranslationPolling(OperationResourcePolling):
+    """Implements a Location polling."""
+
+    def can_poll(self, pipeline_response: PipelineResponseType) -> bool:
+        """Answer if this polling method could be used.
+
+        :param pipeline_response: The PipelineResponse type
+        :type pipeline_response: PipelineResponseType
+        :return: Whether polling should be performed.
+        :rtype: bool
+        """
+        response = pipeline_response.http_response
+        can_poll = self._operation_location_header in response.headers
+        if can_poll:
+            return True
+
+        if not _is_empty(response):
+            body = _as_json(response)
+            status = body.get("status")
+            if status:
+                return True
+        return False
+
+    def _set_async_url_if_present(self, response: ResponseType) -> None:
+        location_header = response.headers.get(self._operation_location_header)
+        if location_header:
+            self._async_url = location_header
+        else:
+            self._async_url = response.request.url
+
+    def get_status(self, pipeline_response: PipelineResponseType) -> str:
+        """Process the latest status update retrieved from a 'location' header.
+
+        :param azure.core.pipeline.PipelineResponse pipeline_response: latest REST call response.
+        :return: The current operation status
+        :rtype: str
+        :raises: BadResponse if response has no body and not status 202.
+        """
+        response = pipeline_response.http_response
+        if not _is_empty(response):
+            body = _as_json(response)
+            status = body.get("status")
+            if status:
+                return self._map_nonstandard_statuses(status, body)
+            raise BadResponse("No status found in body")
+        raise BadResponse("The response from long running operation does not contain a body.")
+
+    def _map_nonstandard_statuses(self, status: str, body: Dict[str, Any]) -> str:
+        """Map non-standard statuses.
+
+        :param str status: lro process status.
+        :param str body: pipeline response body.
+        :return: The current operation status.
+        :rtype: str
+        """
+        if status == "ValidationFailed":
+            self.raise_error(body)
+        return status
+
+    def raise_error(self, body: Dict[str, Any]) -> None:
+        error = body["error"]
+        if body["error"].get("innerError", None):
+            error = body["error"]["innerError"]
+        http_response_error = HttpResponseError(message="({}): {}".format(error["code"], error["message"]))
+        http_response_error.error = ODataV4Format(error)  # set error.code
+        raise http_response_error
 
 
 class DocumentTranslationApiVersion(str, Enum, metaclass=CaseInsensitiveEnumMeta):
@@ -51,37 +283,123 @@ def validate_api_version(api_version: str) -> None:
             )
         ) from exc
 
-import json
-import datetime
-from typing import Any, List, Union, TYPE_CHECKING, overload, Optional, cast
-from azure.core.tracing.decorator import distributed_trace
-from azure.core.paging import ItemPaged
-from azure.core.credentials import AzureKeyCredential
-from ._client import (
-    DocumentTranslationClient as _BatchDocumentTranslationClient,
-)
-from .models import StartTranslationDetails
-from .models._patch import (
-    TranslationStatus,
-    DocumentStatus,
-    DocumentTranslationInput,
-    DocumentTranslationFileFormat,
-    convert_status,
-    StorageInputType,
-    TranslationGlossary
-)
-from ._user_agent import USER_AGENT
-from ._polling import TranslationPolling, DocumentTranslationLROPollingMethod, DocumentTranslationLROPoller
-#from ._api_version import DocumentTranslationApiVersion
-if TYPE_CHECKING:
-    from azure.core.credentials import TokenCredential
+
+
+def get_translation_input(args, kwargs, continuation_token):
+    try:
+        inputs = kwargs.pop("inputs", None)
+        if not inputs:
+            inputs = args[0]
+        request = (
+            DocumentTranslationInput._to_generated_list(inputs)  # pylint: disable=protected-access
+            if not continuation_token
+            else None
+        )
+    except (AttributeError, TypeError, IndexError):
+        try:
+            source_url = kwargs.pop("source_url", None)
+            if not source_url:
+                source_url = args[0]
+            target_url = kwargs.pop("target_url", None)
+            if not target_url:
+                target_url = args[1]
+            target_language = kwargs.pop("target_language", None)
+            if not target_language:
+                target_language = args[2]
+
+            # Additional kwargs
+            source_language = kwargs.pop("source_language", None)
+            prefix = kwargs.pop("prefix", None)
+            suffix = kwargs.pop("suffix", None)
+            storage_type = kwargs.pop("storage_type", None)
+            category_id = kwargs.pop("category_id", None)
+            glossaries = kwargs.pop("glossaries", None)
+
+            request = [
+                _BatchRequest(
+                    source=_SourceInput(
+                        source_url=source_url,
+                        filter=_DocumentFilter(prefix=prefix, suffix=suffix),
+                        language=source_language,
+                    ),
+                    targets=[
+                        _TargetInput(
+                            target_url=target_url,
+                            language=target_language,
+                            glossaries=[g._to_generated() for g in glossaries]  # pylint: disable=protected-access
+                            if glossaries
+                            else None,
+                            category=category_id,
+                        )
+                    ],
+                    storage_type=storage_type,
+                )
+            ]
+        except (AttributeError, TypeError, IndexError) as exc:
+            raise ValueError(
+                "Pass 'inputs' for multiple inputs or 'source_url', 'target_url', "
+                "and 'target_language' for a single input."
+            ) from exc
+
+    return request
+
+
+def get_http_logging_policy(**kwargs):
+    http_logging_policy = HttpLoggingPolicy(**kwargs)
+    http_logging_policy.allowed_header_names.update(
+        {
+            "Operation-Location",
+            "Content-Encoding",
+            "Vary",
+            "apim-request-id",
+            "X-RequestId",
+            "Set-Cookie",
+            "X-Powered-By",
+            "Strict-Transport-Security",
+            "x-content-type-options",
+        }
+    )
+    http_logging_policy.allowed_query_params.update(
+        {
+            "$top",
+            "$skip",
+            "$maxpagesize",
+            "ids",
+            "statuses",
+            "createdDateTimeUtcStart",
+            "createdDateTimeUtcEnd",
+            "$orderBy",
+        }
+    )
+    return http_logging_policy
+
+
+def convert_datetime(date_time: Union[str, datetime.datetime]) -> datetime.datetime:
+    if isinstance(date_time, datetime.datetime):
+        return date_time
+    if isinstance(date_time, str):
+        try:
+            return datetime.datetime.strptime(date_time, "%Y-%m-%d")
+        except ValueError:
+            try:
+                return datetime.datetime.strptime(date_time, "%Y-%m-%dT%H:%M:%SZ")
+            except ValueError:
+                return datetime.datetime.strptime(date_time, "%Y-%m-%d %H:%M:%S")
+    raise TypeError("Bad datetime type")
+
+
+def convert_order_by(order_by: Optional[List[str]]) -> Optional[List[str]]:
+    if order_by:
+        order_by = [order.replace("created_on", "createdDateTimeUtc") for order in order_by]
+    return order_by
+
 
 
 class DocumentTranslationClient:
     def __init__(
         self,
         endpoint: str,
-        credential: Union[AzureKeyCredential, "TokenCredential"],
+        credential: Union[AzureKeyCredential, TokenCredential],
         **kwargs: Any
     ) -> None:
         """DocumentTranslationClient is your interface to the Document Translation service.
@@ -122,17 +440,17 @@ class DocumentTranslationClient:
             raise ValueError("Parameter 'endpoint' must be a string.") from exc
         self._credential = credential
         self._api_version = kwargs.pop("api_version", None)
-        if hasattr(self._api_version, "value"):
-            self._api_version = cast(DocumentTranslationApiVersion, self._api_version)
-            self._api_version = self._api_version.value
-        authentication_policy = get_authentication_policy(credential)
+        # TODO how to support API version - v1.0 vs date-based?
+        # if hasattr(self._api_version, "value"):
+        #     self._api_version = cast(DocumentTranslationApiVersion, self._api_version)
+        #     self._api_version = self._api_version.value
         polling_interval = kwargs.pop("polling_interval", POLLING_INTERVAL)
+        
+        from ._client import DocumentTranslationClient as _BatchDocumentTranslationClient
         self._client = _BatchDocumentTranslationClient(
             endpoint=self._endpoint,
-            credential=credential,  
-            api_version=self._api_version,
-            sdk_moniker=USER_AGENT,
-            authentication_policy=kwargs.pop("authentication_policy", authentication_policy),
+            credential=credential,
+            # api_version=self._api_version,
             http_logging_policy=kwargs.pop("http_logging_policy", get_http_logging_policy()),
             polling_interval=polling_interval,
             **kwargs
@@ -311,15 +629,15 @@ class DocumentTranslationClient:
 
         pipeline_response = None
         if continuation_token:
-            pipeline_response = self._client.document_translation.get_translation_status(
+            pipeline_response = self._client.get_translation_status(
                 continuation_token,
                 cls=lambda pipeline_response, _, response_headers: pipeline_response,
             )
 
         callback = kwargs.pop("cls", deserialization_callback)
         return cast(DocumentTranslationLROPoller[ItemPaged[DocumentStatus]],
-            self._client.document_translation.begin_start_translation(
-                body=StartTranslationDetails(inputs=inputs),
+            self._client.begin_start_translation(
+                body=_StartTranslationDetails(inputs=inputs),
                 polling=DocumentTranslationLROPollingMethod(
                     timeout=polling_interval,
                     lro_algorithms=[TranslationPolling()],
@@ -345,11 +663,11 @@ class DocumentTranslationClient:
         :raises ~azure.core.exceptions.HttpResponseError or ~azure.core.exceptions.ResourceNotFoundError:
         """
 
-        translation_status = self._client.document_translation.get_translation_status(
+        translation_status = self._client.get_translation_status(
             translation_id, **kwargs
         )
         return TranslationStatus._from_generated(  # pylint: disable=protected-access
-            translation_status
+            _TranslationStatus(translation_status)
         )
 
     @distributed_trace
@@ -366,7 +684,7 @@ class DocumentTranslationClient:
         :raises ~azure.core.exceptions.HttpResponseError or ~azure.core.exceptions.ResourceNotFoundError:
         """
 
-        self._client.document_translation.cancel_translation(translation_id, **kwargs)
+        self._client.cancel_translation(translation_id, **kwargs)
 
     @distributed_trace
     def list_translation_statuses(
@@ -422,7 +740,7 @@ class DocumentTranslationClient:
             generated_model,
         ):  # pylint: disable=protected-access
             return TranslationStatus._from_generated(
-                generated_model
+                _TranslationStatus(generated_model)
             )  # pylint: disable=protected-access
 
         model_conversion_function = kwargs.pop(
@@ -433,7 +751,7 @@ class DocumentTranslationClient:
         )
 
         return cast(ItemPaged[TranslationStatus],
-            self._client.document_translation.get_translations_status(
+            self._client.get_translations_status(
                 cls=model_conversion_function,
                 created_date_time_utc_start=created_after,
                 created_date_time_utc_end=created_before,
@@ -504,7 +822,7 @@ class DocumentTranslationClient:
 
         def _convert_from_generated_model(generated_model):
             return DocumentStatus._from_generated(  # pylint: disable=protected-access
-                generated_model
+                _DocumentStatus(generated_model)
             )
 
         model_conversion_function = kwargs.pop(
@@ -515,7 +833,7 @@ class DocumentTranslationClient:
         )
 
         return cast(ItemPaged[DocumentStatus],
-            self._client.document_translation.get_documents_status(
+            self._client.get_documents_status(
                 id=translation_id,
                 cls=model_conversion_function,
                 created_date_time_utc_start=created_after,
@@ -524,7 +842,7 @@ class DocumentTranslationClient:
                 order_by=order_by,
                 statuses=statuses,
                 top=top,
-                skip=skip,  
+                skip=skip,
                 **kwargs
             )
         )
@@ -540,11 +858,11 @@ class DocumentTranslationClient:
         :raises ~azure.core.exceptions.HttpResponseError or ~azure.core.exceptions.ResourceNotFoundError:
         """
 
-        document_status = self._client.document_translation.get_document_status(
+        document_status = self._client.get_document_status(
             translation_id, document_id, **kwargs
         )
         return DocumentStatus._from_generated(  # pylint: disable=protected-access
-            document_status
+            _DocumentStatus(document_status)
         )
 
     @distributed_trace
@@ -557,7 +875,10 @@ class DocumentTranslationClient:
         """
 
         glossary_formats = (
-            self._client.document_translation.get_supported_glossary_formats(**kwargs)
+            self._client.get_supported_formats(
+                type="glossary",
+                **kwargs
+            )
         )
         return DocumentTranslationFileFormat._from_generated_list(  # pylint: disable=protected-access
             glossary_formats.value
@@ -573,154 +894,34 @@ class DocumentTranslationClient:
         """
 
         document_formats = (
-            self._client.document_translation.get_supported_document_formats(**kwargs)
+            self._client.get_supported_formats(
+                type="document",
+                **kwargs
+            )
         )
         return DocumentTranslationFileFormat._from_generated_list(  # pylint: disable=protected-access
             document_formats.value
         )
 
-# ------------------------------------
-# Copyright (c) Microsoft Corporation.
-# Licensed under the MIT License.
-# ------------------------------------
-
-import datetime
-from typing import Union, Optional, List
-from azure.core.credentials import AzureKeyCredential
-from azure.core.pipeline.policies import AzureKeyCredentialPolicy
-from azure.core.pipeline.policies import HttpLoggingPolicy
-from azure.ai.translation.document.models import (
-    BatchRequest as _BatchRequest,
-    SourceInput as _SourceInput,
-    TargetInput as _TargetInput,
-    DocumentFilter as _DocumentFilter    
-)
-
-from azure.ai.translation.document.models._patch import DocumentTranslationInput
-
-COGNITIVE_KEY_HEADER = "Ocp-Apim-Subscription-Key"
-POLLING_INTERVAL = 1
+__all__: List[str] = [
+    "DocumentTranslationClient",
+    "DocumentTranslationApiVersion",
+    "DocumentTranslationLROPoller",
+    # re-export models at this level for backwards compatibility
+    "TranslationGlossary",
+    "TranslationTarget",
+    "DocumentTranslationInput",
+    "TranslationStatus",
+    "DocumentStatus",
+    "DocumentTranslationError",
+    "DocumentTranslationFileFormat",
+]  # Add all objects you want publicly available to users at this package level
 
 
-def get_translation_input(args, kwargs, continuation_token):
-    try:
-        inputs = kwargs.pop("inputs", None)
-        if not inputs:
-            inputs = args[0]
-        request = (
-            DocumentTranslationInput._to_generated_list(inputs)  # pylint: disable=protected-access
-            if not continuation_token
-            else None
-        )
-    except (AttributeError, TypeError, IndexError):
-        try:
-            source_url = kwargs.pop("source_url", None)
-            if not source_url:
-                source_url = args[0]
-            target_url = kwargs.pop("target_url", None)
-            if not target_url:
-                target_url = args[1]
-            target_language = kwargs.pop("target_language", None)
-            if not target_language:
-                target_language = args[2]
+def patch_sdk():
+    """Do not remove from this file.
 
-            # Additional kwargs
-            source_language = kwargs.pop("source_language", None)
-            prefix = kwargs.pop("prefix", None)
-            suffix = kwargs.pop("suffix", None)
-            storage_type = kwargs.pop("storage_type", None)
-            category_id = kwargs.pop("category_id", None)
-            glossaries = kwargs.pop("glossaries", None)
-
-            request = [
-                _BatchRequest(
-                    source=_SourceInput(
-                        source_url=source_url,
-                        filter=_DocumentFilter(prefix=prefix, suffix=suffix),
-                        language=source_language,
-                    ),
-                    targets=[
-                        _TargetInput(
-                            target_url=target_url,
-                            language=target_language,
-                            glossaries=[g._to_generated() for g in glossaries]  # pylint: disable=protected-access
-                            if glossaries
-                            else None,
-                            category=category_id,
-                        )
-                    ],
-                    storage_type=storage_type,
-                )
-            ]
-        except (AttributeError, TypeError, IndexError) as exc:
-            raise ValueError(
-                "Pass 'inputs' for multiple inputs or 'source_url', 'target_url', "
-                "and 'target_language' for a single input."
-            ) from exc
-
-    return request
-
-
-def get_authentication_policy(credential):
-    authentication_policy = None
-    if credential is None:
-        raise ValueError("Parameter 'credential' must not be None.")
-    if isinstance(credential, AzureKeyCredential):
-        authentication_policy = AzureKeyCredentialPolicy(name=COGNITIVE_KEY_HEADER, credential=credential)
-    elif credential is not None and not hasattr(credential, "get_token"):
-        raise TypeError(
-            "Unsupported credential: {}. Use an instance of AzureKeyCredential "
-            "or a token credential from azure.identity".format(type(credential))
-        )
-
-    return authentication_policy
-
-
-def get_http_logging_policy(**kwargs):
-    http_logging_policy = HttpLoggingPolicy(**kwargs)
-    http_logging_policy.allowed_header_names.update(
-        {
-            "Operation-Location",
-            "Content-Encoding",
-            "Vary",
-            "apim-request-id",
-            "X-RequestId",
-            "Set-Cookie",
-            "X-Powered-By",
-            "Strict-Transport-Security",
-            "x-content-type-options",
-        }
-    )
-    http_logging_policy.allowed_query_params.update(
-        {
-            "$top",
-            "$skip",
-            "$maxpagesize",
-            "ids",
-            "statuses",
-            "createdDateTimeUtcStart",
-            "createdDateTimeUtcEnd",
-            "$orderBy",
-        }
-    )
-    return http_logging_policy
-
-
-def convert_datetime(date_time: Union[str, datetime.datetime]) -> datetime.datetime:
-    if isinstance(date_time, datetime.datetime):
-        return date_time
-    if isinstance(date_time, str):
-        try:
-            return datetime.datetime.strptime(date_time, "%Y-%m-%d")
-        except ValueError:
-            try:
-                return datetime.datetime.strptime(date_time, "%Y-%m-%dT%H:%M:%SZ")
-            except ValueError:
-                return datetime.datetime.strptime(date_time, "%Y-%m-%d %H:%M:%S")
-    raise TypeError("Bad datetime type")
-
-
-def convert_order_by(order_by: Optional[List[str]]) -> Optional[List[str]]:
-    if order_by:
-        order_by = [order.replace("created_on", "createdDateTimeUtc") for order in order_by]
-    return order_by
+    `patch_sdk` is a last resort escape hatch that allows you to do customizations
+    you can't accomplish using the techniques described in
+    https://aka.ms/azsdk/python/dpcodegen/python/customize
+    """
