--- sdk/translation/azure-ai-translation-document/azure/ai/translation/document/_user_agent.py
@@ -1,8 +1,0 @@
-# ------------------------------------
-# Copyright (c) Microsoft Corporation.
-# Licensed under the MIT License.
-# ------------------------------------
-
-from ._version import VERSION
-
-USER_AGENT = f"ai-translation-document/{VERSION}"
