+++ sdk/translation/azure-ai-translation-document/pyproject.toml
@@ -1,3 +1,6 @@
 [tool.azure-sdk-build]
 pyright = false
 strict_sphinx = true
+
+[tool.generate]
+autorest-post-process = true
