+++ sdk/translation/azure-ai-translation-document/tests/conftest.py
@@ -17,11 +17,6 @@ from devtools_testutils import (
 )
 from azure.storage.blob import BlobServiceClient
 
-# Ignore async tests for Python < 3.5
-collect_ignore_glob = []
-if sys.version_info < (3, 5):
-    collect_ignore_glob.append("*_async.py")
-
 
 @pytest.fixture(scope="session", autouse=True)
 def add_sanitizers(test_proxy):
