+++ sdk/translation/azure-ai-translation-document/tests/test_supported_formats.py
@@ -13,7 +13,7 @@ from devtools_testutils import recorded_by_proxy
 from azure.ai.translation.document import DocumentTranslationClient
 
 DocumentTranslationClientPreparer = functools.partial(_DocumentTranslationClientPreparer, DocumentTranslationClient)
-from azure.ai.translation.document.models._enums import FormatType
+
 
 class TestSupportedFormats(DocumentTranslationTest):
     @DocumentTranslationPreparer()
@@ -22,10 +22,11 @@ class TestSupportedFormats(DocumentTranslationTest):
     def test_supported_document_formats(self, **kwargs):
         client = kwargs.pop("client")
         # get supported formats
-        supported_doc_formats = client.get_supported_formats(type=FormatType.DOCUMENT)
+        supported_doc_formats = client.get_supported_document_formats()
+        print(supported_doc_formats)
         assert supported_doc_formats is not None
         # validate
-        for doc_format in supported_doc_formats.value:
+        for doc_format in supported_doc_formats:
             self._validate_format(doc_format)
 
     @DocumentTranslationPreparer()
@@ -34,8 +35,8 @@ class TestSupportedFormats(DocumentTranslationTest):
     def test_supported_glossary_formats(self, **kwargs):
         client = kwargs.pop("client")
         # get supported formats
-        supported_glossary_formats = client.get_supported_formats(type=FormatType.GLOSSARY)
+        supported_glossary_formats = client.get_supported_glossary_formats()
         assert supported_glossary_formats is not None
         # validate
-        for glossary_format in supported_glossary_formats.value:
+        for glossary_format in supported_glossary_formats:
             self._validate_format(glossary_format)
