+++ sdk/translation/azure-ai-translation-document/tests/test_supported_formats_async.py
@@ -13,7 +13,7 @@ from devtools_testutils.aio import recorded_by_proxy_async
 from azure.ai.translation.document.aio import DocumentTranslationClient
 
 DocumentTranslationClientPreparer = functools.partial(_DocumentTranslationClientPreparer, DocumentTranslationClient)
-from azure.ai.translation.document.models._enums import FormatType
+
 
 class TestSupportedFormats(AsyncDocumentTranslationTest):
     @DocumentTranslationPreparer()
@@ -22,10 +22,10 @@ class TestSupportedFormats(AsyncDocumentTranslationTest):
     async def test_supported_document_formats(self, **kwargs):
         client = kwargs.pop("client")
         # get supported formats
-        supported_doc_formats = await client.get_supported_formats(type=FormatType.DOCUMENT)
+        supported_doc_formats = await client.get_supported_document_formats()
         assert supported_doc_formats is not None
         # validate
-        for doc_format in supported_doc_formats.value:
+        for doc_format in supported_doc_formats:
             self._validate_format(doc_format)
 
     @DocumentTranslationPreparer()
@@ -34,8 +34,8 @@ class TestSupportedFormats(AsyncDocumentTranslationTest):
     async def test_supported_glossary_formats(self, **kwargs):
         client = kwargs.pop("client")
         # get supported formats
-        supported_glossary_formats = await client.get_supported_formats(type=FormatType.GLOSSARY)
+        supported_glossary_formats = await client.get_supported_glossary_formats()
         assert supported_glossary_formats is not None
         # validate
-        for glossary_format in supported_glossary_formats.value:
+        for glossary_format in supported_glossary_formats:
             self._validate_format(glossary_format)
